Name that Character Game
This is a trivia game for guessing the names of popular cartoon characters from the 90s and 2000s. There are two levels, multiple choice (four options) or typing the name of the character in the textbox. There is a skip question button if you are stumped. You are given a point for each correct answer There is a timer counting down and you must get as many points as possible before the timer runs out. Once the timer runs out, the game is over and a stats page with the score is displayed.